Change run to script in Snakemake rules
