+++ 2_process.smk
@@ -2,15 +2,6 @@ import os
 import numpy as np
 import pandas as pd
 
-import sys
-sys.path.insert(1, './2_process/src')
-from process_utils import make_lake_metadata_augmented
-from process_utils import make_obs_interpolated
-from process_utils import unzip_all
-
-# Include is necessary to call get_mntoha_input_files
-# This include is redundant unless this file is referenced directly rather than through the main Snakefile
-include: "1_fetch.smk"
 configfile: "2_process/process_config.yaml"
 
 
@@ -18,17 +9,15 @@ def mntoha_lake_sequence_files(wildcards):
     """
     List all MNTOHA lake sequence .npy files for training and testing.
 
-    :param wildcards: Snakemake wildcards (unused).
+    :param wildcards: Snakemake wildcards (empty).
     :returns: List of lake training/testing sequence files.
 
     """
     # make this function dependent on fetch_all
     # needed because lake_metadata.csv is used to determine lake_sequence_files
-    ck_output = checkpoints.fetch_all.get(**wildcards).output[0]
-    out_dir = os.path.join(config["process_mntoha"]["out_dir"], 'mntoha_sequences')
-    lake_metadata_file = config["process_mntoha"]["metadata_file"]
+    lake_metadata_file = checkpoints.fetch_mntoha_metadata.get(**wildcards).output[0]
     lake_metadata = pd.read_csv(lake_metadata_file)
-    lake_sequence_files = [os.path.join(out_dir, f'sequences_{site_id}.npy') for site_id in lake_metadata.site_id]
+    lake_sequence_files = [os.path.join('2_process/out/mntoha_sequences', f'sequences_{site_id}.npy') for site_id in lake_metadata.site_id]
     return lake_sequence_files
 
 
@@ -75,11 +64,30 @@ rule mntoha_lake_sequences:
 # Add elevation to MNTOHA lake metadata
 rule augment_mntoha_lake_metadata:
     input:
-        "1_fetch/out/downloaded_files.txt"
+        "1_fetch/out/lake_metadata.csv"
     output:
-        config["process_mntoha"]["metadata_augmented_file"]
-    run:
-        make_lake_metadata_augmented(config["process_mntoha"]["metadata_file"], output[0])
+        "2_process/tmp/mntoha/lake_metadata_augmented.csv"
+    script:
+        "2_process/src/make_lake_metadata_augmented.py"
+
+
+def mntoha_obs_file(wildcards):
+    '''
+    Return MNTOHA temperature observations filepath.
+
+    Depend on unzip_archive checkpoint to ensure that
+    temperature_observations.csv gets unzipped.
+
+    :param wildcards: Snakemake wildcards (empty).
+    :returns: Path of MNTOHA temperature observations csv
+
+    '''
+    # Trigger checkpoint to unzip temperature_observations.csv
+    obs_file_directory = checkpoints.unzip_archive.get(
+        file_category='obs_mntoha',
+        directory_name='temperature_observations'
+    ).output[0]
+    return os.path.join(obs_file_directory, "temperature_observations.csv")
 
 
 # Add column of observation depths interpolated to nearest modeling mesh node
@@ -88,6 +96,8 @@ rule interpolate_mntoha_obs_depths:
         config["process_mntoha"]["unzip_log_file"],
     output:
         config["process_mntoha"]["obs_interpolated_file"]
-    run:
-        make_obs_interpolated(config["process_mntoha"]["obs_file"], output[0], config["process_mntoha"]["depths"])
+    params:
+        depths=config["process_mntoha"]["depths"]
+    script:
+        "2_process/src/make_obs_interpolated.py"
